backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiofiles==23.2.1
python-dotenv==1.0.1
pandas==2.2.0
xlsxwriter==3.1.9
openai==1.12.0
python-magic==0.4.27
Pillow==10.2.0
pytesseract==0.3.10
email-validator==2.1.0.post1
psycopg2-binary==2.9.9
alembic==1.12.1
pytest==7.4.3
httpx==0.25.1
python-dateutil==2.8.2
boto3==1.29.3 